V.5
So far I have been able to provide a good amount of responsiveness for
the website using some jQuery built in animations. Talked to the client
and they felt happy with the visual element.

TO-DO's:

* Use the SimplyBook API to show customer reviews on the site.

* Smooth out the scroll when clicking on a navbar element

* Get contact bar on bottom of site filled out.

* Acquire pictures for site

* Include a lotus flower.... Make it a quick ole .svg

* Write content in text areas.

^
|
|
Lots of work to do still.

Please enter the commit message for your changes. Lines starting